Continue authorization

+++ backend/routes/posts.js
@@ -36,7 +36,8 @@ router.post('', checkAuth, multer({ storage: storage }).single('image'), (req, r
   const post = new Post({
     title: req.body.title,
     content: req.body.content,
-    imagePath: url + '/images/' + req.file.filename
+    imagePath: url + '/images/' + req.file.filename,
+    creator: req.userData.userId
   });
   post.save().then(result => {
     res.status(200).json({
@@ -63,18 +64,22 @@ router.put('/:id', checkAuth, multer({ storage: storage }).single('image'), (req
     content: req.body.content,
     imagePath: imagePath
   });
-  Post.updateOne({ _id: req.params.id }, post).then(result => {
-    console.log(result);
-    res.status(200).json({ message: 'Updated successfull!' });
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+    if (result.nModified > 0) {
+      res.status(200).json({ message: 'Updated successfull!' });
+    } else {
+      res.status(401).json({ message: 'Not Authorized!' });
+    }
   });
 });
 
 router.delete('/:id', checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then(result => {
-    console.log(result);
-    res.status(200).json({
-      message: 'Post deleted!'
-    });
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
+    if (result.n > 0) {
+      res.status(200).json({ message: 'Deletion successfull!' });
+    } else {
+      res.status(401).json({ message: 'Not Authorized!' });
+    }
   });
 });
